@@ -4,15 +4,15 @@
   "private": true,
   "homepage": "https://hnngo.github.io/portfolio",
   "dependencies": {
-    "classnames": "^2.2.6",
-    "node-sass": "^4.13.1",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.0.4",
-    "react-tooltip": "4.1.2",
-    "redux": "^4.0.0",
-    "redux-promise": "^0.6.0"
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/material": "^5.11.1",
+    "classnames": "^2.3.2",
+    "node-sass": "^7.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "react-tooltip": "^5.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -32,7 +32,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "gh-pages": "^2.0.1",
+    "gh-pages": "^4.0.0",
     "ws": "3.3.2"
   }
 }
