@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from "react";
 
 // Components
 import SectionTitle from "../../shared/SectionTitle";
@@ -12,40 +12,45 @@ import cx from "classnames";
 
 import styles from "./style.module.scss";
 
-export default class ExperienceSection extends Component {
-  constructor(props) {
-    super(props);
+const EXP_DATA = EXP_BOARD_DATA.filter(e => e.show === "on");
 
-    this.state = {
-      isInited: false,
-      resizeEvent: undefined,
-      eConHeight: window.innerHeight,
-      ebWidtth: 0,
-      offset: -70,
-      transOffset: 60,
-      boardScale: [],
-      boardOpacity: [],
-      boardZindex: [],
-      boardTransX: [],
-      boardTransXOrigin: [],
-    };
-  }
+const ExperienceSection = ({ show }) => {
+  const [isInited, setIsInited] = React.useState(false);
+  const [eConHeight, setEConHeight] = React.useState(window.innerHeight);
+  const [ebWidtth, setEbWidtth] = React.useState(0);
+  const [boardScale, setBoardScale] = React.useState([]);
+  const [boardOpacity, setBoardOpacity] = React.useState([]);
+  const [boardZindex, setBoardZindex] = React.useState([]);
+  const [boardTransX, setBoardTransX] = React.useState([]);
+  const [boardTransXOrigin, setBoardTransXOrigin] = React.useState([]);
 
-  componentDidUpdate() {
-    if (this.props.show && !this.state.isInited) {
+  const initResizeEvent = React.useCallback(() => {
+    setInterval(() => {
+      let currentHeight = window.innerHeight;
+
+      if (currentHeight !== eConHeight) {
+        setEConHeight(currentHeight);
+      }
+    }, 300);
+  }, [eConHeight]);
+
+  React.useEffect(() => {
+    if (show && !isInited) {
       // Setup resize event checking
-      this.resizeEvent();
+      initResizeEvent();
 
       // Initial setup style
-      let boardScale = EXP_BOARD_DATA.map((_, i) => 1 - i * 0.1);
-      let boardOpacity = EXP_BOARD_DATA.map((_, i) => (i === 0 ? 1 : 0.3));
-      let boardZindex = EXP_BOARD_DATA.map((_, i) => EXP_BOARD_DATA.length - i);
+      let newBoardScale = EXP_DATA.map((_, i) => 1 - i * 0.1);
+      let newBoardOpacity = EXP_DATA.map((_, i) => (i === 0 ? 1 : 0.3));
+      let newBoardZindex = EXP_DATA.map(
+        (_, i) => EXP_DATA.length - i
+      );
 
       // Update transX
       const qBoardWidth = document.querySelector(
         "#" + SECTIONS_ID.EXP_BOARD + "0"
       ).clientWidth;
-      let boardTransX = boardScale.map((item, i) => {
+      let newBoardTransX = newBoardScale.map((item, i) => {
         // let realSize = qBoardWidth * (+item);
         let deltaSize = (qBoardWidth * (1 - +item)) / 2;
         let currentWidth = window.innerWidth;
@@ -55,166 +60,130 @@ export default class ExperienceSection extends Component {
         return -(deltaSize + i * newTransOffset);
       });
 
-      this.setState({
-        isInited: true,
-        ebWidtth: qBoardWidth,
-        prevCovertedSelect: 0,
-        boardScale,
-        boardOpacity,
-        boardZindex,
-        boardTransX,
-        boardTransXOrigin: boardTransX,
-      });
+      setIsInited(true);
+      setEbWidtth(qBoardWidth);
+      setBoardScale(newBoardScale);
+      setBoardOpacity(newBoardOpacity);
+      setBoardZindex(newBoardZindex);
+      setBoardTransX(newBoardTransX);
+      setBoardTransXOrigin(newBoardTransX);
     }
-  }
-
-  resizeEvent() {
-    const resizeEvent = setInterval(() => {
-      let currentHeight = window.innerHeight;
-
-      if (currentHeight !== this.state.eConHeight) {
-        this.setState({
-          eConHeight: currentHeight,
-        });
-      }
-    }, 300);
-
-    this.setState({ resizeEvent });
-  }
+  }, [show, isInited, initResizeEvent, ebWidtth]);
 
   // Handle slide animation
-  handleSelectMileStone(selectedVal) {
-    // Convert the data follow the direction 0 > max
-    let numberOfData = EXP_BOARD_DATA.length - 1;
-    let convertedVal = (selectedVal / 100) * -numberOfData + numberOfData;
+  const handleSelectMileStone = React.useCallback(
+    (selectedVal) => {
+      // Convert the data follow the direction 0 > max
+      let numberOfData = EXP_DATA.length - 1;
+      let convertedVal = (selectedVal / 100) * -numberOfData + numberOfData;
 
-    let boardScale = this.state.boardScale;
-    let boardOpacity = this.state.boardOpacity;
-    let boardZindex = this.state.boardZindex;
-    let boardTransX = [...this.state.boardTransX];
-    let boardTransXOrigin = this.state.boardTransXOrigin;
+      let newBoardScale = boardScale;
+      let newBoardOpacity = boardOpacity;
+      let newBoardZindex = boardZindex;
+      let newBoardTransX = [...boardTransX];
+      let newBoardTransXOrigin = boardTransXOrigin;
 
-    EXP_BOARD_DATA.forEach((_, i) => {
-      let delta = Math.abs(i - convertedVal);
-      let signedDelta = i - convertedVal;
-
-      // Update opacity, zIndex
-      let deltaOpacity = 1 - 0.1;
-      let deltaZindex = 5 - 4;
-      let deltaTransX;
-
-      if (delta < 1) {
-        boardOpacity[i] = 1 - delta * deltaOpacity;
-        boardZindex[i] = Math.round(5 - delta * deltaZindex);
-
-        deltaTransX = boardTransXOrigin[1];
-      } else if (signedDelta < 0) {
-        boardOpacity[i] = 0.3;
-        boardZindex[i] = 1;
-
-        deltaTransX =
-          boardTransXOrigin[Math.floor(convertedVal) - i] -
-          boardTransXOrigin[Math.floor(convertedVal) - i - 1];
-      } else {
-        boardOpacity[i] = 0.3;
-        boardZindex[i] = 1;
-
-        deltaTransX = boardTransXOrigin[i] - boardTransXOrigin[i - 1];
+      // Ignore if the boards not initialized
+      if (newBoardScale.length !== EXP_DATA.length) {
+        return;
       }
 
-      // Update scale
-      boardScale[i] = 1 - delta * 0.1;
+      EXP_DATA.forEach((_, i) => {
+        let delta = Math.abs(i - convertedVal);
+        let signedDelta = i - convertedVal;
 
-      // Update transX
-      boardTransX[i] = deltaTransX * signedDelta;
-    });
+        // Update opacity, zIndex
+        let deltaOpacity = 1 - 0.1;
+        let deltaZindex = 5 - 4;
+        let deltaTransX;
 
-    this.setState({
-      boardScale,
-      boardOpacity,
-      boardZindex,
-      boardTransX,
-      prevCovertedSelect: convertedVal,
-    });
-  }
+        
+        if (delta < 1) {
+          newBoardOpacity[i] = 1 - delta * deltaOpacity;
+          newBoardZindex[i] = Math.round(5 - delta * deltaZindex);
 
-  renderExpBoard() {
-    return EXP_BOARD_DATA.map((item, i) => (
-      <ExpBoard
-        key={i}
-        keyBoard={i}
-        employer={item.employer}
-        designation={item.designation}
-        compLogo={item.compLogo}
-        headSentence={item.headSentence}
-        jDHeader={item.jDHeader}
-        jobDescription={item.jobDescription}
-        location={item.location}
-        timeFrom={item.timeFrom}
-        timeTo={item.timeTo}
-        scale={this.state.boardScale[i]}
-        opacity={this.state.boardOpacity[i]}
-        zIndex={this.state.boardZindex[i]}
-        transX={this.state.boardTransX[i]}
-      />
-    ));
-  }
+          deltaTransX = newBoardTransXOrigin[1];
+        } else if (signedDelta < 0) {
+          newBoardOpacity[i] = 0.3;
+          newBoardZindex[i] = 1;
 
-  render() {
-    return (
-      <div
-        id={SECTIONS_ID.EXPERIENCE}
-        className={styles.container}
-        style={{
-          height: this.state.eConHeight,
-        }}
-      >
-        {!this.props.show ? null : (
-          <div
-            id={SECTIONS_ID.EXP_CONTAINER}
-            className={cx(styles.boardContainer, "container")}
-          >
-            <SectionTitle
-              title="Experience"
-              icon="fas fa-history"
-              isDarkTheme={true}
-            />
-            <div>
-              {EXP_BOARD_DATA.map((item, i) => (
-                <ExpBoard
-                  key={i}
-                  keyBoard={i}
-                  employer={item.employer}
-                  designation={item.designation}
-                  compLogo={item.compLogo}
-                  headSentence={item.headSentence}
-                  jDHeader={item.jDHeader}
-                  jobDescription={item.jobDescription}
-                  location={item.location}
-                  timeFrom={item.timeFrom}
-                  timeTo={item.timeTo}
-                  scale={this.state.boardScale[i]}
-                  opacity={this.state.boardOpacity[i]}
-                  zIndex={this.state.boardZindex[i]}
-                  transX={this.state.boardTransX[i]}
-                />
-              ))}
-            </div>
-            <div
-              className={cx(
-                styles.yearSlideContainer,
-                "animated slideInUp slow"
-              )}
-            >
-              <YearSlider
-                onSelectMS={(selectedVal) =>
-                  this.handleSelectMileStone(selectedVal)
-                }
+          deltaTransX =
+            newBoardTransXOrigin[Math.floor(convertedVal) - i] -
+            newBoardTransXOrigin[Math.floor(convertedVal) - i - 1];
+        } else {
+          boardOpacity[i] = 0.3;
+          boardZindex[i] = 1;
+
+          deltaTransX = newBoardTransXOrigin[i] - newBoardTransXOrigin[i - 1];
+        }
+
+        // Update scale
+        newBoardScale[i] = 1 - delta * 0.1;
+
+        // Update transX
+        newBoardTransX[i] = deltaTransX * signedDelta;
+      });
+
+      setBoardScale(newBoardScale);
+      setBoardOpacity(newBoardOpacity);
+      setBoardZindex(newBoardZindex);
+      setBoardTransX(newBoardTransX);
+      setBoardTransXOrigin(newBoardTransXOrigin);
+    },
+    [boardOpacity, boardScale, boardZindex, boardTransX, boardTransXOrigin]
+  );
+
+  return (
+    <div
+      id={SECTIONS_ID.EXPERIENCE}
+      className={styles.container}
+      style={{
+        height: eConHeight,
+      }}
+    >
+      {!show ? null : (
+        <div
+          id={SECTIONS_ID.EXP_CONTAINER}
+          className={cx(styles.boardContainer, "container")}
+        >
+          <SectionTitle
+            title="Experience"
+            icon="fas fa-history"
+            isDarkTheme={true}
+          />
+          <div>
+            {EXP_DATA.map((item, i) => (
+              <ExpBoard
+                key={i}
+                keyBoard={i}
+                employer={item.employer}
+                designation={item.designation}
+                compLogo={item.compLogo}
+                headSentence={item.headSentence}
+                jDHeader={item.jDHeader}
+                jobDescription={item.jobDescription}
+                location={item.location}
+                timeFrom={item.timeFrom}
+                timeTo={item.timeTo}
+                scale={boardScale[i]}
+                opacity={boardOpacity[i]}
+                zIndex={boardZindex[i]}
+                transX={boardTransX[i]}
               />
-            </div>
+            ))}
           </div>
-        )}
-      </div>
-    );
-  }
-}
+          <div
+            className={cx(styles.yearSlideContainer, "animated slideInUp slow")}
+          >
+            <YearSlider
+              expData={[...EXP_DATA].reverse()}
+              onSelectMS={(selectedVal) => handleSelectMileStone(selectedVal)}
+            />
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default ExperienceSection;
