@@ -4,12 +4,11 @@ import React from "react";
 import LazyImage from "../../../../shared/LazyImage";
 import devicePlaceholder from "../../../../style/img/device_ph.png";
 import porjectThumbnailPlaceholder from "../../../../style/img/project_thumbnail_ph.png";
+import Tooltip from "@mui/material/Tooltip";
 
 // Constants and utils
 import cx from "classnames";
 import { PROJECT_INFO, TECH_LOGO } from "../../../../data";
-import ReactTooltip from "react-tooltip";
-import { SECTIONS_ID } from "../../../../shared/constants";
 
 import styles from "./style.module.scss";
 
@@ -19,7 +18,7 @@ const ProjectDetail = ({
   onClickThumbnail,
   onClickBackToBoardView,
   slideLeftAnimation,
-  slideRightAnimation
+  slideRightAnimation,
 }) => {
   const renderProjectDetail = () => {
     if (selectedIndex !== undefined) {
@@ -52,19 +51,20 @@ const ProjectDetail = ({
                 ))}
               </ul>
               <p className={styles.detailTech}>Technologies/Libraries</p>
-              {board.technologies.map((item, i) => {
-                return (
-                  <div key={i} className={styles.detailTechLogoWrapper}>
-                    <LazyImage
-                      className={styles.detailTechLogo}
-                      src={TECH_LOGO[item]}
-                      data-tip={item}
-                      data-for={SECTIONS_ID.TECH_TOOLTIP + i}
-                    />
-                    <ReactTooltip id={SECTIONS_ID.TECH_TOOLTIP + i} />
-                  </div>
-                );
-              })}
+              <div className={styles.detailTechLogoContainer}>
+                {board.technologies.map((item, i) => {
+                  return (
+                    <Tooltip title={item} placement="top" arrow>
+                      <div key={i} className={styles.detailTechLogoWrapper}>
+                        <LazyImage
+                          className={styles.detailTechLogo}
+                          src={TECH_LOGO[item]}
+                        />
+                      </div>
+                    </Tooltip>
+                  );
+                })}
+              </div>
               <div className={styles.detailView}>
                 {board.website.length > 0 && (
                   <a
