@@ -1,4 +1,4 @@
-import * as ABOUT_CONTENT from "./about.json";
-import * as INTRODUCTION from "./introduction.json";
+import ABOUT_CONTENT from "./about.json";
+import INTRODUCTION from "./introduction.json";
 
 export { ABOUT_CONTENT, INTRODUCTION };
