@@ -3,11 +3,11 @@ import cx from "classnames";
 
 import styles from "./style.module.scss";
 
-function LazyImage({ src, alt, placeholder, className, ...rest }) {
+function LazyImage({ src, alt, placeholder, className, id, ...rest }) {
   const [isLoaded, setIsLoaded] = React.useState(null);
 
   return (
-    <>
+    <span id={id}>
       {(isLoaded === null || isLoaded === true) && (
         <img
           className={cx(className, !isLoaded && styles.invisible)}
@@ -22,7 +22,7 @@ function LazyImage({ src, alt, placeholder, className, ...rest }) {
         !!placeholder && (
           <img className={className} src={placeholder} alt={alt} {...rest} />
         )}
-    </>
+    </span>
   );
 }
 
